folder restructure
move marker, circle and etc. to overlays folder
