+++ lib/src/types/interface/index.ts
@@ -2,12 +2,9 @@ export * from './map-options.interface';
 export * from './pixel.interface';
 export * from './size.interface';
 export * from './lng-lat.interface';
-export * from './marker-options.interface';
 export * from './icon.interface';
 export * from './label.interface';
 export * from './info-window-options.interface';
-export * from './polyline-options.interface';
-export * from './polygon-options.interface';
 export * from './toolbar-options.interface';
 export * from './marker-cluster-options.interface';
 export * from './cluster-style.interface';
@@ -15,8 +12,11 @@ export * from './geocoder-options.interface';
 export * from './autocomplete-options.interface';
 export * from './place-search-options.interface';
 export * from './callback.type';
-export * from './circle-options.interface';
-export * from './circle-marker-options.interface';
+export * from './overlays/marker-options.interface';
+export * from './overlays/polyline-options.interface';
+export * from './overlays/polygon-options.interface';
+export * from './overlays/circle-options.interface';
+export * from './overlays/circle-marker-options.interface';
 export * from './overlays/text-options.interface';
 export * from './overlays/bezier-curve-options.interface';
 export * from './overlays/ellipse-options.interface';
