+++ lib/src/types/interface/overlays/polygon-options.interface.ts
@@ -1,13 +1,12 @@
-import { Map } from '../class/amap.map';
-import { LngLat } from '../class/amap.lng-lat';
+import { Map } from '../../class/amap.map';
+import { LngLat } from '../../class/amap.lng-lat';
 
-export interface CircleOptions {
+export interface PolygonOptions {
   map?: Map;
   zIndex?: number;
-  center?: LngLat;
+  path?: LngLat[] | LngLat[][];
   bubble?: boolean;
   cursor?: string;
-  radius?: number;
   strokeColor?: string;
   strokeOpacity?: number;
   strokeWeight?: number;
@@ -16,4 +15,5 @@ export interface CircleOptions {
   strokeStyle?: string;
   strokeDasharray?: number[];
   extData?: any;
+  draggable?: boolean;
 }
