+++ src/index.html
@@ -4,11 +4,11 @@
   <meta charset="utf-8">
   <title>ngx-amap demo</title>
   <base href="/">
-
+	<meta name="keyword" content="Angular,Gaode,Map,ngx,AMap">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
-<body>
+<body class="app header-fixed">
   <app-root></app-root>
 </body>
 </html>
